ai-python/requirements.txt
fastapi
uvicorn
python-multipart

torch
torchvision
torchaudio

transformers
ultralytics
openai-whisper
spacy
geopy
opencv-python
numpy
